+++ CHANGELOG.md
@@ -7,6 +7,17 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [2.3.1] - 2023-12-23
+
+### Added
+
+- Add a few utility methods to `ProgressStats`.
+
+### Changed
+
+- `pyo3` is no longer needed to use this crate as Rust-only library.
+- Updated Rust dependencies.
+
 ## [2.3.0] - 2023-11-05
 
 ### Added
@@ -277,6 +288,7 @@ Full changes: <https://github.com/Decompollaborate/mapfile_parser/compare/702a73
 - Initial release
 
 [unreleased]: https://github.com/Decompollaborate/mapfile_parser/compare/master...develop
+[2.3.1]: https://github.com/Decompollaborate/mapfile_parser/compare/2.3.0...2.3.1
 [2.3.0]: https://github.com/Decompollaborate/mapfile_parser/compare/2.2.1...2.3.0
 [2.2.1]: https://github.com/Decompollaborate/mapfile_parser/compare/2.2.0...2.2.1
 [2.2.0]: https://github.com/Decompollaborate/mapfile_parser/compare/2.1.5...2.2.0